+++ ChangeLog.md
@@ -28,6 +28,7 @@
 - BUGFIX: #5395 Fixed loading trigger condition if it starts with the NOT keyword.
 - BUGFIX: #5398 Fixed Smart Executor for queries, where two joined, unaliased tables return column of the same name, but different case-wise.
 - BUGFIX: #5406 Tracking calls to load_extension() SQL function and loading the same extensions in the Results Counting connection.
+- BUGFIX: #5411 Data View sorting is reset upon modifying table structure.
 - BUGFIX: Fixed arbitrary crashes when opening SQL Editor, while having non-thread-safe SQLite extensions loaded in the active database.
 
 ### 3.4.17